Move the MakeOp to OpCodeTable class, for reduce the use of ConcurrentDictionary
